+++ com.microej.demo.widget/src/main/java/com/microej/demo/widget/selectablelist/widget/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright 2015-2021 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ */
+
+/**
+ * Contains the selectable list page associated widgets.
+ */
+@ej.annotation.NonNullByDefault
+package com.microej.demo.widget.selectablelist.widget;