+++ com.microej.demo.widget/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 7.3.0 - 2021-08-03
+
+### Added
+
+- Add SelectableList page and AnimatedChoice widget.
+
+### Fixed
+
+- Fix launchers to use latest STM32F7508 platform on GitHub.
+
 ## 7.2.0 - 2021-05-10
 
 ### Added